+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Changed
 - The debugger now has better defaults for the list of compiled items ([#4038](https://github.com/julia-vscode/julia-vscode/pull/4038))
+- Users will now only get one modal dialogue if installation of required juliaup channels or julia is cancelled ([#4039](https://github.com/julia-vscode/julia-vscode/pull/4039))
 
 ### Fixed
 - Stopped printing deprecation warnings when recursively setting modules as interpreted or compiled ([DebugAdapter.jl#115](https://github.com/julia-vscode/DebugAdapter.jl/pull/115))